Update start code
Add two stage heating to prevent unwanted nozzle ooze without the
downside of waiting the full length of time needed to heat the hot end
from room temperature.

Remove M83 from the end of start code. Setting the extrusion mode to
relative causes issues with some slicer that does not reset things after
the custom start code. While PrusaSlicer is the intended use case of
this start code, this is a simple enough fix to implement, and it
prevents headache for everyone.